- Ensured History does not overwrite an existing file without prompting.    Added a corresponding test.
 - Prevented History save dialog from filling in the name of the current
   class file as the default filename.

 - Eliminated the use of null as a value in History.writeToFile.

 - Updated the OS X application bundle to use Java 1.4.1 if it is
   available.  It should still work with 1.3.1 if 1.4.1 is not installed.
   (Slight issue: the icon switches to the default Java icon during
   startup in 1.4.1.  It seems some other applications have this problem
   too.)

 - Switched resetInteractionsClasspath to put the paths in config on the
   classpath before the source roots of the open documents.  This is not
   ideal, but it's consistent with the state of the classpath before
   clicking Reset.  (Config is added on startup, and documents are
   appended to the end as they are opened.)  Ideally, we'd like to insert
   the source roots at the front as they are opened, but that's not really
   an option.  This partially alleviates bug 702733, in which the order of
   the Interactions classpath can cause problems with static fields.

 - Made the MainFrameTest silent by using a TestBeep.


git-svn-id: file:///tmp/test-svn/trunk@1290 fe72c1cf-3628-48e9-8b72-1c46755d3cff

+++ drjava/src/edu/rice/cs/drjava/model/repl/History.java
@@ -201,13 +201,23 @@ public class History implements OptionConstants {
   }
   
   /**
-   * Writes this History to a the file selected in the FileSaveSelector
+   * Writes this (unedited) History to the file selected in the FileSaveSelector
    * @param selector File to save to
-   * @param editedVersion If this string is non-null, it will be saved to file
-   * instead of the lines saved in the history. The saved file will still include
+   */
+  public void writeToFile(FileSaveSelector selector) throws IOException {
+    writeToFile(selector, getHistoryAsStringWithSemicolons());
+  }
+  
+  /**
+   * Writes this History to the file selected in the FileSaveSelector
+   * @param selector File to save to
+   * @param editedVersion The edited version of the text to be saved.
+   * The saved file will still include
    * any tags or extensions needed to recognize it as a saved interactions file.
    */
-  public void writeToFile(FileSaveSelector selector, String editedVersion) throws IOException {
+  public void writeToFile(FileSaveSelector selector, String editedVersion) 
+    throws IOException
+  {
     
     File c = null;
     
@@ -217,19 +227,20 @@ public class History implements OptionConstants {
     catch (OperationCanceledException oce) {
       return; // don't need to do anything
     }
+    
+    // Make sure we ask before overwriting
     if (c != null) {
-      if (c.getName().indexOf('.') == -1) {
-        c = new File(c.getAbsolutePath() + "." +
-                     InteractionsHistoryFilter.HIST_EXTENSION);
+      if (!c.exists() || selector.verifyOverwrite()) {
+        if (c.getName().indexOf('.') == -1) {
+          c = new File(c.getAbsolutePath() + "." +
+                       InteractionsHistoryFilter.HIST_EXTENSION);
+        }
+        FileOutputStream fos = new FileOutputStream(c);
+        OutputStreamWriter osw = new OutputStreamWriter(fos);
+        BufferedWriter bw = new BufferedWriter(osw);
+        bw.write(HISTORY_FORMAT_VERSION_2 + editedVersion, 0, HISTORY_FORMAT_VERSION_2.length() + editedVersion.length());
+        bw.close();
       }
-      FileOutputStream fos = new FileOutputStream(c);
-      OutputStreamWriter osw = new OutputStreamWriter(fos);
-      BufferedWriter bw = new BufferedWriter(osw);
-      if (editedVersion == null) {
-        editedVersion = getHistoryAsStringWithSemicolons();
-      }
-      bw.write(HISTORY_FORMAT_VERSION_2 + editedVersion, 0, HISTORY_FORMAT_VERSION_2.length() + editedVersion.length());
-      bw.close();
     }
   }
